update post api

+++ src/pages/Home/Home.jsx
@@ -5,31 +5,57 @@ import {
   useGetAnnouncements,
   useDeleteAnnouncement,
 } from "../../hooks/useAnnouncement";
-import { useGetPosts, useDeletePost } from "../../hooks/usePost";
+import {
+  useGetPostsByReligion,
+  useDeletePost,
+  useUpdatePost,
+} from "../../hooks/usePost";
 import { UserContext } from "../../contexts/UserContext";
+import { ReligionContext } from "../../contexts/ReligionContext";
 import { useNavigate } from "react-router-dom";
 
 function Home() {
   const navigate = useNavigate();
   const { user } = useContext(UserContext);
+  const { religion } = useContext(ReligionContext);
   const { data: announcements } = useGetAnnouncements();
   const { mutate: deleteAnnouncement } = useDeleteAnnouncement();
-  const { data: posts } = useGetPosts();
+  const { data: posts } = useGetPostsByReligion(religion);
   const { mutate: deletePost } = useDeletePost();
+  const { mutate: updatePost } = useUpdatePost();
 
   const onAnnouncementDelete = (id) => {
-    deleteAnnouncement(id);
+    if (window.confirm("คุณต้องการลบประกาศนี้หรือไม่?")) {
+      deleteAnnouncement(id);
+    }
   };
 
   const onPostDelete = (id) => {
-    deletePost(id);
+    if (window.confirm("คุณต้องการลบกระทู้นี้หรือไม่?")) {
+      deletePost(id);
+    }
+  };
+
+  const onHide = (item) => {
+    if (window.confirm("คุณต้องการซ่อนกระทู้นี้หรือไม่?")) {
+      let newItem = { ...item };
+      newItem.isHide = true;
+      updatePost(newItem);
+    }
   };
 
   return (
     <HomeContainer>
       <HomeItem>
         <div className="item-header">Announcement</div>
-        <div className="item-content">
+        <div
+          className="item-content"
+          style={
+            announcements?.length % 2 === 0
+              ? { backgroundColor: "#c4c4c4" }
+              : { backgroundColor: "#fff" }
+          }
+        >
           {announcements?.map((announcement, index) => {
             return (
               <div key={announcement.id}>
@@ -49,22 +75,38 @@ function Home() {
       </HomeItem>
       <HomeItem>
         <div className="item-header">กระทู้ธรรม</div>
-        <div className="item-content">
-          {posts?.map((post, index) => {
-            return (
-              <div key={post.id}>
-                <PostItem
-                  bg={index % 2 === 0 ? "#fff" : "#c4c4c4"}
-                  item={post}
-                  role={user?.role}
-                  userId={user?.id}
-                  type="post"
-                  onDelete={() => onPostDelete(post.id)}
-                  onEdit={() => navigate(`/edit-post/${post.id}`)}
-                />
-              </div>
-            );
-          })}
+        <div
+          className="item-content"
+          style={
+            posts?.filter((post) => {
+              return post.isHide === false;
+            }).length %
+              2 ===
+            0
+              ? { backgroundColor: "#c4c4c4" }
+              : { backgroundColor: "#fff" }
+          }
+        >
+          {posts
+            ?.filter((post) => {
+              return post.isHide === false;
+            })
+            .map((post, index) => {
+              return (
+                <div key={post.id}>
+                  <PostItem
+                    bg={index % 2 === 0 ? "#fff" : "#c4c4c4"}
+                    item={post}
+                    role={user?.role}
+                    userId={user?.id}
+                    type="post"
+                    onDelete={() => onPostDelete(post.id)}
+                    onEdit={() => navigate(`/edit-post/${post.id}`)}
+                    onHide={() => onHide(post)}
+                  />
+                </div>
+              );
+            })}
         </div>
       </HomeItem>
     </HomeContainer>
